move logic to build new network state message
